LPD-89767 Extract Info Panel tests to info-panel directory
